+++ awscli/customizations/s3/subcommands.py
@@ -30,7 +30,7 @@ from awscli.customizations.s3.utils import find_bucket_key, uni_print, \
     AppendFilter, find_dest_path_comp_key, human_readable_size
 from awscli.customizations.s3.syncstrategy.base import MissingFileSync, \
     SizeAndLastModifiedSync, NeverSync
-
+from awscli.customizations.s3 import transferconfig
 
 
 RECURSIVE = {'name': 'recursive', 'action': 'store_true', 'dest': 'dir_op',
@@ -38,25 +38,33 @@ RECURSIVE = {'name': 'recursive', 'action': 'store_true', 'dest': 'dir_op',
                  "Command is performed on all files or objects "
                  "under the specified directory or prefix.")}
 
+
 HUMAN_READABLE = {'name': 'human-readable', 'action': 'store_true',
                   'help_text': "Displays file sizes in human readable format."}
 
-SUMMARIZE = {'name': 'summarize', 'action': 'store_true', 'help_text': (
-                 "Displays summary information (number of objects, total size).")}
+
+SUMMARIZE = {'name': 'summarize', 'action': 'store_true',
+             'help_text': (
+                 "Displays summary information "
+                 "(number of objects, total size).")}
+
 
 DRYRUN = {'name': 'dryrun', 'action': 'store_true',
           'help_text': (
               "Displays the operations that would be performed using the "
               "specified command without actually running them.")}
 
+
 QUIET = {'name': 'quiet', 'action': 'store_true',
          'help_text': (
              "Does not display the operations performed from the specified "
              "command.")}
 
+
 FORCE = {'name': 'force', 'action': 'store_true',
          'help_text': (
              "Deletes all objects in the bucket including the bucket itself.")}
+
 
 FOLLOW_SYMLINKS = {'name': 'follow-symlinks', 'action': 'store_true',
                    'default': True, 'group_name': 'follow_symlinks',
@@ -69,9 +77,11 @@ FOLLOW_SYMLINKS = {'name': 'follow-symlinks', 'action': 'store_true',
                        "nor ``--no-follow-symlinks`` is specifed, the default "
                        "is to follow symlinks.")}
 
+
 NO_FOLLOW_SYMLINKS = {'name': 'no-follow-symlinks', 'action': 'store_false',
                       'dest': 'follow_symlinks', 'default': True,
                       'group_name': 'follow_symlinks'}
+
 
 NO_GUESS_MIME_TYPE = {'name': 'no-guess-mime-type', 'action': 'store_false',
                       'dest': 'guess_mime_type', 'default': True,
@@ -80,10 +90,12 @@ NO_GUESS_MIME_TYPE = {'name': 'no-guess-mime-type', 'action': 'store_false',
                           "uploaded files.  By default the mime type of a "
                           "file is guessed when it is uploaded.")}
 
+
 CONTENT_TYPE = {'name': 'content-type', 'nargs': 1,
                 'help_text': (
                     "Specify an explicit content type for this operation.  "
                     "This value overrides any guessed mime types.")}
+
 
 EXCLUDE = {'name': 'exclude', 'action': AppendFilter, 'nargs': 1,
            'dest': 'filters',
@@ -91,11 +103,13 @@ EXCLUDE = {'name': 'exclude', 'action': AppendFilter, 'nargs': 1,
                "Exclude all files or objects from the command that matches "
                "the specified pattern.")}
 
+
 INCLUDE = {'name': 'include', 'action': AppendFilter, 'nargs': 1,
            'dest': 'filters',
            'help_text': (
                "Don't exclude files or objects "
                "in the command that match the specified pattern")}
+
 
 ACL = {'name': 'acl', 'nargs': 1,
        'choices': ['private', 'public-read', 'public-read-write',
@@ -108,31 +122,37 @@ ACL = {'name': 'acl', 'nargs': 1,
            "``bucket-owner-read``, ``bucket-owner-full-control`` and "
            "``log-delivery-write``.")}
 
-GRANTS = {'name': 'grants', 'nargs': '+',
-          'help_text': (
-              "Grant specific permissions to individual users or groups. You "
-              "can supply a list of grants of the form::<p/>  --grants "
-              "Permission=Grantee_Type=Grantee_ID [Permission=Grantee_Type="
-              "Grantee_ID ...]<p/>Each value contains the following elements:"
-              "<p/><ul><li><code>Permission</code> - Specifies "
-              "the granted permissions, and can be set to read, readacl, "
-              "writeacl, or full.</li><li><code>Grantee_Type</code> - "
-              "Specifies how the grantee is to be identified, and can be set "
-              "to uri, emailaddress, or id.</li><li><code>Grantee_ID</code> - "
-              "Specifies the grantee based on Grantee_Type.</li></ul>The "
-              "<code>Grantee_ID</code> value can be one of:<ul><li><b>uri</b> "
-              "- The group's URI. For more information, see "
-              '<a href="http://docs.aws.amazon.com/AmazonS3/latest/dev/ACLOverview.html#SpecifyingGrantee">'
-              "Who Is a Grantee?</a></li>"
-              "<li><b>emailaddress</b> - The account's email address.</li>"
-              "<li><b>id</b> - The account's canonical ID</li></ul>"
-              "</li></ul>"
-              "For more information on Amazon S3 access control, see "
-              '<a href="http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAuthAccess.html">Access Control</a>')}
+
+GRANTS = {
+    'name': 'grants', 'nargs': '+',
+    'help_text': (
+        'Grant specific permissions to individual users or groups. You '
+        'can supply a list of grants of the form::<p/>  --grants '
+        'Permission=Grantee_Type=Grantee_ID [Permission=Grantee_Type='
+        'Grantee_ID ...]<p/>Each value contains the following elements:'
+        '<p/><ul><li><code>Permission</code> - Specifies '
+        'the granted permissions, and can be set to read, readacl, '
+        'writeacl, or full.</li><li><code>Grantee_Type</code> - '
+        'Specifies how the grantee is to be identified, and can be set '
+        'to uri, emailaddress, or id.</li><li><code>Grantee_ID</code> - '
+        'Specifies the grantee based on Grantee_Type.</li></ul>The '
+        '<code>Grantee_ID</code> value can be one of:<ul><li><b>uri</b> '
+        '- The group\'s URI. For more information, see '
+        '<a href="http://docs.aws.amazon.com/AmazonS3/latest/dev/'
+        'ACLOverview.html#SpecifyingGrantee">'
+        'Who Is a Grantee?</a></li>'
+        '<li><b>emailaddress</b> - The account\'s email address.</li>'
+        '<li><b>id</b> - The account\'s canonical ID</li></ul>'
+        '</li></ul>'
+        'For more information on Amazon S3 access control, see '
+        '<a href="http://docs.aws.amazon.com/AmazonS3/latest/dev/'
+        'UsingAuthAccess.html">Access Control</a>')}
+
 
 SSE = {'name': 'sse', 'action': 'store_true',
        'help_text': (
            "Enable Server Side Encryption of the object in S3")}
+
 
 STORAGE_CLASS = {'name': 'storage-class', 'nargs': 1,
                  'choices': ['STANDARD', 'REDUCED_REDUNDANCY'],
@@ -140,6 +160,7 @@ STORAGE_CLASS = {'name': 'storage-class', 'nargs': 1,
                      "The type of storage to use for the object. "
                      "Valid choices are: STANDARD | REDUCED_REDUNDANCY. "
                      "Defaults to 'STANDARD'")}
+
 
 WEBSITE_REDIRECT = {'name': 'website-redirect', 'nargs': 1,
                     'help_text': (
@@ -149,15 +170,18 @@ WEBSITE_REDIRECT = {'name': 'website-redirect', 'nargs': 1,
                         "stores the value of this header in the object "
                         "metadata.")}
 
+
 CACHE_CONTROL = {'name': 'cache-control', 'nargs': 1,
                  'help_text': (
                      "Specifies caching behavior along the "
                      "request/reply chain.")}
 
+
 CONTENT_DISPOSITION = {'name': 'content-disposition', 'nargs': 1,
                        'help_text': (
                            "Specifies presentational information "
                            "for the object.")}
+
 
 CONTENT_ENCODING = {'name': 'content-encoding', 'nargs': 1,
                     'help_text': (
@@ -166,8 +190,10 @@ CONTENT_ENCODING = {'name': 'content-encoding', 'nargs': 1,
                         "mechanisms must be applied to obtain the media-type "
                         "referenced by the Content-Type header field.")}
 
+
 CONTENT_LANGUAGE = {'name': 'content-language', 'nargs': 1,
                     'help_text': ("The language the content is in.")}
+
 
 SOURCE_REGION = {'name': 'source-region', 'nargs': 1,
                  'help_text': (
@@ -179,8 +205,13 @@ SOURCE_REGION = {'name': 'source-region', 'nargs': 1,
                      "specified the region of the source will be the same "
                      "as the region of the destination bucket.")}
 
-EXPIRES = {'name': 'expires', 'nargs': 1, 'help_text': ("The date and time at "
-           "which the object is no longer cacheable.")}
+
+EXPIRES = {
+    'name': 'expires', 'nargs': 1,
+    'help_text': (
+        "The date and time at which the object is no longer cacheable.")
+}
+
 
 INDEX_DOCUMENT = {'name': 'index-document',
                   'help_text': (
@@ -192,15 +223,18 @@ INDEX_DOCUMENT = {'name': 'index-document',
                       'images/index.html) The suffix must not be empty and '
                       'must not include a slash character.')}
 
+
 ERROR_DOCUMENT = {'name': 'error-document',
                   'help_text': (
                       'The object key name to use when '
                       'a 4XX class error occurs.')}
 
+
 ONLY_SHOW_ERRORS = {'name': 'only-show-errors', 'action': 'store_true',
                     'help_text': (
                         'Only errors and warnings are displayed. All other '
                         'output is suppressed.')}
+
 
 EXPECTED_SIZE = {'name': 'expected-size',
                  'help_text': (
@@ -213,10 +247,10 @@ EXPECTED_SIZE = {'name': 'expected-size',
 
 
 PAGE_SIZE = {'name': 'page-size', 'cli_type_name': 'integer',
-			 'help_text': (
-				 'The number of results to return in each response to a list '
-				 'operation. The default value is 1000 (the maximum allowed). '
-				 'Using a lower value may help if an operation times out.')}
+             'help_text': (
+                 'The number of results to return in each response to a list '
+                 'operation. The default value is 1000 (the maximum allowed). '
+                 'Using a lower value may help if an operation times out.')}
 
 
 TRANSFER_ARGS = [DRYRUN, QUIET, RECURSIVE, INCLUDE, EXCLUDE, ACL,
@@ -355,7 +389,7 @@ class ListCommand(S3Command):
                         str(last_mod.day).zfill(2),
                         str(last_mod.hour).zfill(2),
                         str(last_mod.minute).zfill(2),
-                       str(last_mod.second).zfill(2))
+                        str(last_mod.second).zfill(2))
         last_mod_str = "%s-%s-%s %s:%s:%s" % last_mod_tup
         return last_mod_str.ljust(19, ' ')
 
@@ -363,7 +397,10 @@ class ListCommand(S3Command):
         """
         This function creates the size string when objects are being listed.
         """
-        size_str = human_readable_size(size) if self._human_readable else str(size)
+        if self._human_readable:
+            size_str = human_readable_size(size)
+        else:
+            size_str = str(size)
         return size_str.rjust(10, ' ')
 
     def _print_summary(self):
@@ -372,7 +409,10 @@ class ListCommand(S3Command):
         """
         print_str = str(self._total_objects)
         uni_print("\nTotal Objects: ".rjust(15, ' ') + print_str + "\n")
-        print_str = human_readable_size(self._size_accumulator) if self._human_readable else str(self._size_accumulator)
+        if self._human_readable:
+            print_str = human_readable_size(self._size_accumulator)
+        else:
+            print_str = str(self._size_accumulator)
         uni_print("Total Size: ".rjust(15, ' ') + print_str + "\n")
 
 
@@ -421,16 +461,19 @@ class S3TransferCommand(S3Command):
         super(S3TransferCommand, self)._run_main(parsed_args, parsed_globals)
         self._convert_path_args(parsed_args)
         params = self._build_call_parameters(parsed_args, {})
-        cmd_params = CommandParameters(self._session, self.NAME, params,
+        cmd_params = CommandParameters(self.NAME, params,
                                        self.USAGE)
         cmd_params.add_region(parsed_globals)
         cmd_params.add_endpoint_url(parsed_globals)
         cmd_params.add_verify_ssl(parsed_globals)
         cmd_params.add_page_size(parsed_args)
         cmd_params.add_paths(parsed_args.paths)
-        cmd_params.check_force(parsed_globals)
+        self._handle_rm_force(parsed_globals, cmd_params.parameters)
+        runtime_config = transferconfig.RuntimeConfig().build_config(
+            **self._session.get_scoped_config().get('s3', {}))
         cmd = CommandArchitecture(self._session, self.NAME,
-                                  cmd_params.parameters)
+                                  cmd_params.parameters,
+                                  runtime_config)
         cmd.set_endpoints()
         cmd.create_instructions()
         return cmd.run()
@@ -454,6 +497,26 @@ class S3TransferCommand(S3Command):
                 enc_path = dec_path.encode('utf-8')
                 new_path = enc_path.decode('utf-8')
                 parsed_args.paths[i] = new_path
+
+    def _handle_rm_force(self, parsed_globals, parameters):
+        """
+        This function recursive deletes objects in a bucket if the force
+        parameters was thrown when using the remove bucket command.
+        """
+        # XXX: This shouldn't really be here.  This was originally moved from
+        # the CommandParameters class to here, but this is still not the ideal
+        # place for this code.  This should be moved
+        # to either the CommandArchitecture class, or the RbCommand class where
+        # the actual operations against S3 are performed.  This may require
+        # some refactoring though to move this to either of those classes.
+        # For now, moving this out of CommandParameters allows for that class
+        # to be kept simple.
+        if 'force' in parameters:
+            if parameters['force']:
+                bucket = find_bucket_key(parameters['src'][5:])[0]
+                path = 's3://' + bucket
+                del_objects = RmCommand(self._session)
+                del_objects([path, '--recursive'], parsed_globals)
 
 
 class CpCommand(S3TransferCommand):
@@ -526,11 +589,12 @@ class CommandArchitecture(object):
     lsit of instructions to wire together an assortment of generators to
     perform the command.
     """
-    def __init__(self, session, cmd, parameters):
+    def __init__(self, session, cmd, parameters, runtime_config=None):
         self.session = session
         self.cmd = cmd
         self.parameters = parameters
         self.instructions = []
+        self._runtime_config = runtime_config
         self._service = self.session.get_service('s3')
         self._endpoint = None
         self._source_endpoint = None
@@ -668,9 +732,11 @@ class CommandArchitecture(object):
                                      service=self._service,
                                      endpoint=self._endpoint,
                                      is_stream=True)]
-        file_info_builder = FileInfoBuilder(self._service, self._endpoint,
-                                 self._source_endpoint, self.parameters)
+        file_info_builder = FileInfoBuilder(
+            self._service, self._endpoint,
+            self._source_endpoint, self.parameters)
         s3handler = S3Handler(self.session, self.parameters,
+                              runtime_config=self._runtime_config,
                               result_queue=result_queue)
         s3_stream_handler = S3StreamHandler(self.session, self.parameters,
                                             result_queue=result_queue)
@@ -689,7 +755,7 @@ class CommandArchitecture(object):
                             's3_handler': [s3handler]}
         elif self.cmd == 'cp' and self.parameters['is_stream']:
             command_dict = {'setup': [stream_file_info],
-                             's3_handler': [s3_stream_handler]}
+                            's3_handler': [s3_stream_handler]}
         elif self.cmd == 'cp':
             command_dict = {'setup': [files],
                             'file_generator': [file_generator],
@@ -749,12 +815,16 @@ class CommandParameters(object):
     This class is used to do some initial error based on the
     parameters and arguments passed to the command line.
     """
-    def __init__(self, session, cmd, parameters, usage):
+    def __init__(self, cmd, parameters, usage):
         """
         Stores command name and parameters.  Ensures that the ``dir_op`` flag
         is true if a certain command is being used.
+
+        :param cmd: The name of the command, e.g. "rm".
+        :param parameters: A dictionary of parameters.
+        :param usage: A usage string
+
         """
-        self.session = session
         self.cmd = cmd
         self.parameters = parameters
         self.usage = usage
@@ -822,14 +892,6 @@ class CommandParameters(object):
                     path += '/'
                     paths[i] = path
 
-    def _verify_bucket_exists(self, bucket_name):
-        session = self.session
-        service = session.get_service('s3')
-        endpoint = service.get_endpoint(self.parameters['region'])
-        operation = service.get_operation('ListObjects')
-        # This will raise an exception if the bucket does not exist.
-        operation.call(endpoint, bucket=bucket_name, max_keys=0)
-
     def check_path_type(self, paths):
         """
         This initial check ensures that the path types for the specified
@@ -881,21 +943,6 @@ class CommandParameters(object):
                     pass
             else:
                 raise Exception("Error: Local path does not exist")
-
-    def check_force(self, parsed_globals):
-        """
-        This function recursive deletes objects in a bucket if the force
-        parameters was thrown when using the remove bucket command.
-        """
-        if 'force' in self.parameters:
-            if self.parameters['force']:
-                bucket = find_bucket_key(self.parameters['src'][5:])[0]
-                path = 's3://' + bucket
-                try:
-                    del_objects = RmCommand(self.session)
-                    del_objects([path, '--recursive'], parsed_globals)
-                except:
-                    pass
 
     def add_region(self, parsed_globals):
         self.parameters['region'] = parsed_globals.region