+++ tests/integration/customizations/s3/test_plugin.py
@@ -33,7 +33,7 @@ from awscli.compat import six
 from awscli.testutils import unittest, FileCreator, get_stdout_encoding
 from awscli.testutils import aws as _aws
 from tests.unit.customizations.s3 import create_bucket as _create_bucket
-from awscli.customizations.s3 import constants
+from awscli.customizations.s3.transferconfig import DEFAULTS
 
 
 @contextlib.contextmanager
@@ -162,12 +162,16 @@ class BaseS3CLICommand(unittest.TestCase):
         return parsed['Buckets']
 
     def content_type_for_key(self, bucket_name, key_name):
+        parsed = self.head_object(bucket_name, key_name)
+        return parsed['ContentType']
+
+    def head_object(self, bucket_name, key_name):
         operation = self.service.get_operation('HeadObject')
         endpoint = self.service.get_endpoint(self.regions[bucket_name])
         http, parsed = operation.call(
             endpoint, bucket=bucket_name, key=key_name)
         self.assertEqual(http.status_code, 200)
-        return parsed['ContentType']
+        return parsed
 
     def assert_no_errors(self, p):
         self.assertEqual(
@@ -496,6 +500,19 @@ class TestCp(BaseS3CLICommand):
         # Assert that the file was downloaded properly.
         with open(local_filename, 'r') as f:
             self.assertEqual(f.read(), contents)
+
+    def test_website_redirect_ignore_paramfile(self):
+        bucket_name = self.create_bucket()
+        foo_txt = self.files.create_file('foo.txt', 'bar')
+        website_redirect = 'http://someserver'
+        p = aws('s3 cp %s s3://%s/foo.txt --website-redirect %s' %
+                (foo_txt, bucket_name, website_redirect))
+        self.assert_no_errors(p)
+
+        # Ensure that the web address is used as opposed to the contents
+        # of the web address. We can check via a head object.
+        response = self.head_object(bucket_name, 'foo.txt')
+        self.assertEqual(response['WebsiteRedirectLocation'], website_redirect)
 
 
 class TestSync(BaseS3CLICommand):
@@ -1177,7 +1194,9 @@ class TestDryrun(BaseS3CLICommand):
 class TestMemoryUtilization(BaseS3CLICommand):
     # These tests verify the memory utilization and growth are what we expect.
     def extra_setup(self):
-        expected_memory_usage = constants.NUM_THREADS * constants.CHUNKSIZE
+        self.num_threads = DEFAULTS['max_concurrent_requests']
+        self.chunk_size = DEFAULTS['multipart_chunksize']
+        expected_memory_usage = self.num_threads * self.chunk_size
         # margin for things like python VM overhead, botocore service
         # objects, etc.  1.5 is really generous, perhaps over time this can be
         # lowered.
@@ -1234,7 +1253,7 @@ class TestMemoryUtilization(BaseS3CLICommand):
         # is increased by two chunksizes because that is the maximum
         # amount of chunks that will be queued while not being operated on
         # by a thread when performing a streaming multipart upload.
-        max_mem_allowed = self.max_mem_allowed + 2 * constants.CHUNKSIZE
+        max_mem_allowed = self.max_mem_allowed + 2 * self.chunk_size
 
         full_command = 's3 cp - s3://%s/foo.txt' % bucket_name
         with open(foo_txt, 'rb') as f: